refactor: split optimizer folding
