+++ src/optimize/fold/casts.rs
@@ -1,0 +1,81 @@
+use super::super::*;
+use super::scalar::{scalar_value, ScalarValue};
+
+pub(super) fn try_fold_cast(target: &CastType, expr: &Expr) -> Option<ExprKind> {
+    let value = scalar_value(expr)?;
+    match target {
+        CastType::Int => try_fold_cast_int(value),
+        CastType::Float => try_fold_cast_float(value),
+        CastType::String => try_fold_cast_string(value),
+        CastType::Bool => Some(ExprKind::BoolLiteral(value.truthy())),
+        CastType::Array => None,
+    }
+}
+
+fn try_fold_cast_int(value: ScalarValue) -> Option<ExprKind> {
+    match value {
+        ScalarValue::Null => Some(ExprKind::IntLiteral(0)),
+        ScalarValue::Bool(value) => Some(ExprKind::IntLiteral(i64::from(value))),
+        ScalarValue::Int(value) => Some(ExprKind::IntLiteral(value)),
+        ScalarValue::Float(value) => truncate_float_to_i64(value).map(ExprKind::IntLiteral),
+        ScalarValue::String(value) => parse_string_cast_int(&value).map(ExprKind::IntLiteral),
+    }
+}
+
+fn try_fold_cast_float(value: ScalarValue) -> Option<ExprKind> {
+    match value {
+        ScalarValue::Null => Some(ExprKind::FloatLiteral(0.0)),
+        ScalarValue::Bool(value) => Some(ExprKind::FloatLiteral(if value { 1.0 } else { 0.0 })),
+        ScalarValue::Int(value) => Some(ExprKind::FloatLiteral(value as f64)),
+        ScalarValue::Float(value) => Some(ExprKind::FloatLiteral(value)),
+        ScalarValue::String(value) => parse_string_cast_float(&value).map(ExprKind::FloatLiteral),
+    }
+}
+
+fn try_fold_cast_string(value: ScalarValue) -> Option<ExprKind> {
+    match value {
+        ScalarValue::Null => Some(ExprKind::StringLiteral(String::new())),
+        ScalarValue::Bool(value) => Some(ExprKind::StringLiteral(if value {
+            "1".to_string()
+        } else {
+            String::new()
+        })),
+        ScalarValue::Int(value) => Some(ExprKind::StringLiteral(value.to_string())),
+        ScalarValue::Float(_value) => None,
+        ScalarValue::String(value) => Some(ExprKind::StringLiteral(value)),
+    }
+}
+
+fn truncate_float_to_i64(value: f64) -> Option<i64> {
+    if !value.is_finite() {
+        return None;
+    }
+    let truncated = value.trunc();
+    if truncated < i64::MIN as f64 || truncated > i64::MAX as f64 {
+        return None;
+    }
+    Some(truncated as i64)
+}
+
+fn parse_string_cast_int(value: &str) -> Option<i64> {
+    if let Ok(parsed) = value.parse::<i64>() {
+        return Some(parsed);
+    }
+    if let Ok(parsed) = value.parse::<f64>() {
+        return truncate_float_to_i64(parsed);
+    }
+    if value.chars().all(|ch| ch.is_ascii_alphabetic()) {
+        return Some(0);
+    }
+    None
+}
+
+fn parse_string_cast_float(value: &str) -> Option<f64> {
+    if let Ok(parsed) = value.parse::<f64>() {
+        return Some(parsed);
+    }
+    if value.chars().all(|ch| ch.is_ascii_alphabetic()) {
+        return Some(0.0);
+    }
+    None
+}
