+++ src/optimize/fold/ops.rs
@@ -1,0 +1,231 @@
+use super::super::*;
+use super::scalar::{
+    compare_numeric, int_literal, loose_eq, numeric_literal, scalar_value, spaceship_numeric,
+    strict_eq, ScalarValue,
+};
+
+pub(super) fn try_fold_negate(expr: &Expr) -> Option<ExprKind> {
+    match &expr.kind {
+        ExprKind::IntLiteral(value) => value.checked_neg().map(ExprKind::IntLiteral),
+        ExprKind::FloatLiteral(value) => Some(ExprKind::FloatLiteral(-value)),
+        _ => None,
+    }
+}
+
+pub(super) fn try_fold_not(expr: &Expr) -> Option<ExprKind> {
+    Some(ExprKind::BoolLiteral(!scalar_value(expr)?.truthy()))
+}
+
+pub(super) fn try_fold_bit_not(expr: &Expr) -> Option<ExprKind> {
+    match &expr.kind {
+        ExprKind::IntLiteral(value) => Some(ExprKind::IntLiteral(!value)),
+        _ => None,
+    }
+}
+
+pub(super) fn try_fold_binary_op(op: &BinOp, left: &Expr, right: &Expr) -> Option<ExprKind> {
+    match op {
+        BinOp::Concat => try_fold_concat(left, right),
+        BinOp::Add | BinOp::Sub | BinOp::Mul | BinOp::Div | BinOp::Pow => {
+            try_fold_numeric_binop(op, left, right)
+        }
+        BinOp::Mod => try_fold_int_mod(left, right),
+        BinOp::BitAnd | BinOp::BitOr | BinOp::BitXor | BinOp::ShiftLeft | BinOp::ShiftRight => {
+            try_fold_bitwise_binop(op, left, right)
+        }
+        BinOp::And | BinOp::Or | BinOp::Xor => try_fold_logical_binop(op, left, right),
+        BinOp::Eq
+        | BinOp::NotEq
+        | BinOp::StrictEq
+        | BinOp::StrictNotEq
+        | BinOp::Lt
+        | BinOp::Gt
+        | BinOp::LtEq
+        | BinOp::GtEq
+        | BinOp::Spaceship => try_fold_compare_binop(op, left, right),
+        _ => None,
+    }
+}
+
+fn try_fold_concat(left: &Expr, right: &Expr) -> Option<ExprKind> {
+    let ExprKind::StringLiteral(left) = &left.kind else {
+        return None;
+    };
+    let ExprKind::StringLiteral(right) = &right.kind else {
+        return None;
+    };
+    Some(ExprKind::StringLiteral(format!("{left}{right}")))
+}
+
+fn try_fold_numeric_binop(op: &BinOp, left: &Expr, right: &Expr) -> Option<ExprKind> {
+    if let (Some(left), Some(right)) = (int_literal(left), int_literal(right)) {
+        return try_fold_int_numeric_binop(op, left, right);
+    }
+
+    let (left, right) = (numeric_literal(left)?, numeric_literal(right)?);
+    if matches!(op, BinOp::Div) && right == 0.0 {
+        return None;
+    }
+    let result = match op {
+        BinOp::Add => left + right,
+        BinOp::Sub => left - right,
+        BinOp::Mul => left * right,
+        BinOp::Div => left / right,
+        BinOp::Pow => left.powf(right),
+        _ => return None,
+    };
+    if result.is_finite() {
+        Some(ExprKind::FloatLiteral(result))
+    } else {
+        None
+    }
+}
+
+fn try_fold_int_numeric_binop(op: &BinOp, left: i64, right: i64) -> Option<ExprKind> {
+    match op {
+        BinOp::Add => left.checked_add(right).map(ExprKind::IntLiteral),
+        BinOp::Sub => left.checked_sub(right).map(ExprKind::IntLiteral),
+        BinOp::Mul => left.checked_mul(right).map(ExprKind::IntLiteral),
+        BinOp::Div => {
+            if right == 0 {
+                None
+            } else {
+                Some(ExprKind::FloatLiteral(left as f64 / right as f64))
+            }
+        }
+        BinOp::Pow => {
+            let result = (left as f64).powf(right as f64);
+            if result.is_finite() {
+                Some(ExprKind::FloatLiteral(result))
+            } else {
+                None
+            }
+        }
+        _ => None,
+    }
+}
+
+fn try_fold_int_mod(left: &Expr, right: &Expr) -> Option<ExprKind> {
+    let (left, right) = (int_literal(left)?, int_literal(right)?);
+    if right == 0 {
+        None
+    } else {
+        Some(ExprKind::IntLiteral(left % right))
+    }
+}
+
+fn try_fold_bitwise_binop(op: &BinOp, left: &Expr, right: &Expr) -> Option<ExprKind> {
+    let (left, right) = (int_literal(left)?, int_literal(right)?);
+    match op {
+        BinOp::BitAnd => Some(ExprKind::IntLiteral(left & right)),
+        BinOp::BitOr => Some(ExprKind::IntLiteral(left | right)),
+        BinOp::BitXor => Some(ExprKind::IntLiteral(left ^ right)),
+        BinOp::ShiftLeft => {
+            let shift = u32::try_from(right).ok()?;
+            left.checked_shl(shift).map(ExprKind::IntLiteral)
+        }
+        BinOp::ShiftRight => {
+            let shift = u32::try_from(right).ok()?;
+            left.checked_shr(shift).map(ExprKind::IntLiteral)
+        }
+        _ => None,
+    }
+}
+
+fn try_fold_logical_binop(op: &BinOp, left: &Expr, right: &Expr) -> Option<ExprKind> {
+    let left = scalar_value(left)?;
+    let right = scalar_value(right)?;
+    let result = match op {
+        BinOp::And => left.truthy() && right.truthy(),
+        BinOp::Or => left.truthy() || right.truthy(),
+        BinOp::Xor => left.truthy() ^ right.truthy(),
+        _ => return None,
+    };
+    Some(ExprKind::BoolLiteral(result))
+}
+
+fn try_fold_compare_binop(op: &BinOp, left: &Expr, right: &Expr) -> Option<ExprKind> {
+    match op {
+        BinOp::Eq => Some(ExprKind::BoolLiteral(loose_eq(left, right)?)),
+        BinOp::NotEq => Some(ExprKind::BoolLiteral(!loose_eq(left, right)?)),
+        BinOp::StrictEq => Some(ExprKind::BoolLiteral(strict_eq(left, right)?)),
+        BinOp::StrictNotEq => Some(ExprKind::BoolLiteral(!strict_eq(left, right)?)),
+        BinOp::Lt => Some(ExprKind::BoolLiteral(compare_numeric(left, right, |l, r| l < r)?)),
+        BinOp::Gt => Some(ExprKind::BoolLiteral(compare_numeric(left, right, |l, r| l > r)?)),
+        BinOp::LtEq => Some(ExprKind::BoolLiteral(compare_numeric(left, right, |l, r| l <= r)?)),
+        BinOp::GtEq => Some(ExprKind::BoolLiteral(compare_numeric(left, right, |l, r| l >= r)?)),
+        BinOp::Spaceship => Some(ExprKind::IntLiteral(spaceship_numeric(left, right)?)),
+        _ => None,
+    }
+}
+
+pub(super) fn try_fold_null_coalesce(value: &Expr, default: &Expr) -> Option<ExprKind> {
+    let value = scalar_value(value)?;
+    let default = scalar_value(default)?;
+    if matches!(value, ScalarValue::Null) {
+        Some(default.into_expr_kind())
+    } else {
+        Some(value.into_expr_kind())
+    }
+}
+
+pub(super) fn try_fold_ternary(
+    condition: &Expr,
+    then_expr: &Expr,
+    else_expr: &Expr,
+) -> Option<ExprKind> {
+    let condition = scalar_value(condition)?;
+    let then_expr = scalar_value(then_expr)?;
+    let else_expr = scalar_value(else_expr)?;
+    if condition.truthy() {
+        Some(then_expr.into_expr_kind())
+    } else {
+        Some(else_expr.into_expr_kind())
+    }
+}
+
+pub(super) fn try_fold_short_ternary(value: &Expr, default: &Expr) -> Option<ExprKind> {
+    let value = scalar_value(value)?;
+    if value.truthy() {
+        Some(value.into_expr_kind())
+    } else {
+        Some(scalar_value(default)?.into_expr_kind())
+    }
+}
+
+pub(super) fn try_fold_array_access(array: &Expr, index: &Expr) -> Option<ExprKind> {
+    match &array.kind {
+        ExprKind::ArrayLiteral(items) => try_fold_indexed_array_access(items, index),
+        ExprKind::ArrayLiteralAssoc(items) => try_fold_assoc_array_access(items, index),
+        _ => None,
+    }
+}
+
+fn try_fold_indexed_array_access(items: &[Expr], index: &Expr) -> Option<ExprKind> {
+    let ScalarValue::Int(index) = scalar_value(index)? else {
+        return None;
+    };
+    let index = usize::try_from(index).ok()?;
+    let value = items.get(index)?;
+
+    items
+        .iter()
+        .all(|item| scalar_value(item).is_some())
+        .then(|| scalar_value(value).map(ScalarValue::into_expr_kind))
+        .flatten()
+}
+
+fn try_fold_assoc_array_access(items: &[(Expr, Expr)], index: &Expr) -> Option<ExprKind> {
+    let index = scalar_value(index)?;
+    let mut selected = None;
+
+    for (key, value) in items {
+        let key = scalar_value(key)?;
+        let value = scalar_value(value)?;
+        if key == index {
+            selected = Some(value);
+        }
+    }
+
+    selected.map(ScalarValue::into_expr_kind)
+}
